refactor: move wordle logic to separate package
- Created `controller/wordlebot/wordlebot.go` to handle all Wordle game state, validation, and board rendering logic.
- Exposed methods like `IsWordleActive`, `HandleWordleCommand`, and `HandleGuess` for clean external integration.
- Updated `categoryBot` and the main `bot.go` logic to intercept messages/commands using the new package, removing redundant in-file code.
