+++ controller/wordlebot/wordlebot.go
@@ -1,0 +1,231 @@
+package wordlebot
+
+import (
+	"bufio"
+	"fmt"
+	"log"
+	"math/rand"
+	"os"
+	"strings"
+	"sync"
+
+	"github.com/MUSTAFA-A-KHAN/telegram-bot-anime/repository"
+	"github.com/MUSTAFA-A-KHAN/telegram-bot-anime/view"
+	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api"
+	"go.mongodb.org/mongo-driver/mongo"
+)
+
+// WordleState holds the state for a Wordle game in a specific chat.
+type WordleState struct {
+	sync.RWMutex
+	Active      bool
+	Word        string
+	Guesses     []string
+	Attempts    int
+	MaxAttempts int
+}
+
+var (
+	// wordleStates holds the Wordle game state per chat
+	wordleStates = make(map[int64]*WordleState)
+	wordleMutex  = &sync.RWMutex{}
+
+	validWordleWords = make(map[string]bool)
+	wordleWordList   = make([]string, 0)
+	wordsLoaded      bool
+	wordsMutex       sync.RWMutex
+)
+
+// GetOrCreateWordleState safely retrieves or creates a WordleState for a chatID.
+func GetOrCreateWordleState(chatID int64) *WordleState {
+	wordleMutex.Lock()
+	defer wordleMutex.Unlock()
+	if _, exists := wordleStates[chatID]; !exists {
+		wordleStates[chatID] = &WordleState{
+			Guesses:     make([]string, 0),
+			MaxAttempts: 15,
+		}
+	}
+	return wordleStates[chatID]
+}
+
+// LoadWordleWords loads the 5-letter words from words.txt
+func LoadWordleWords() error {
+	wordsMutex.Lock()
+	defer wordsMutex.Unlock()
+
+	if wordsLoaded {
+		return nil
+	}
+
+	// Try to find words.txt starting from current dir or going up
+	paths := []string{
+		"controller/translator/words.txt",
+		"../translator/words.txt",
+		"../../controller/translator/words.txt",
+	}
+
+	var file *os.File
+	var err error
+	for _, p := range paths {
+		file, err = os.Open(p)
+		if err == nil {
+			break
+		}
+	}
+
+	if err != nil {
+		return fmt.Errorf("could not find words.txt: %v", err)
+	}
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		word := strings.TrimSpace(strings.ToLower(scanner.Text()))
+		if len(word) == 5 {
+			validWordleWords[word] = true
+			wordleWordList = append(wordleWordList, word)
+		}
+	}
+
+	if err := scanner.Err(); err != nil {
+		return err
+	}
+
+	wordsLoaded = true
+	log.Printf("Loaded %d Wordle words", len(wordleWordList))
+	return nil
+}
+
+// getRandomWordleWord returns a random 5-letter word
+func getRandomWordleWord() string {
+	wordsMutex.RLock()
+	defer wordsMutex.RUnlock()
+	if len(wordleWordList) == 0 {
+		return "apple" // fallback
+	}
+	return wordleWordList[rand.Intn(len(wordleWordList))]
+}
+
+// validateWordleGuess compares a guess against the target word and returns colored emojis
+func validateWordleGuess(guess, target string) string {
+	guess = strings.ToLower(guess)
+	target = strings.ToLower(target)
+
+	result := make([]string, 5)
+	targetCounts := make(map[rune]int)
+
+	// First pass: count characters in target and check for exact matches (Green)
+	for i, ch := range target {
+		targetCounts[ch]++
+		result[i] = "🟥" // Default to Red
+	}
+
+	// Mark Green
+	for i := 0; i < 5; i++ {
+		if guess[i] == target[i] {
+			result[i] = "🟩"
+			targetCounts[rune(guess[i])]--
+		}
+	}
+
+	// Second pass: check for correct letter in wrong place (Yellow)
+	for i := 0; i < 5; i++ {
+		if guess[i] != target[i] && targetCounts[rune(guess[i])] > 0 {
+			result[i] = "🟨"
+			targetCounts[rune(guess[i])]--
+		}
+	}
+
+	return strings.Join(result, " ")
+}
+
+// buildWordleBoard generates the string representation of the current Wordle board
+func buildWordleBoard(ws *WordleState) string {
+	var sb strings.Builder
+	for _, guess := range ws.Guesses {
+		feedback := validateWordleGuess(guess, ws.Word)
+		sb.WriteString(fmt.Sprintf("%s  %s\n", feedback, strings.ToUpper(guess)))
+	}
+	return sb.String()
+}
+
+// IsWordleActive checks if the Wordle game is active for a given chat ID
+func IsWordleActive(chatID int64) bool {
+	ws := GetOrCreateWordleState(chatID)
+	ws.RLock()
+	defer ws.RUnlock()
+	return ws.Active
+}
+
+// HandleWordleCommand starts a new Wordle game
+func HandleWordleCommand(bot *tgbotapi.BotAPI, chatID int64) {
+	ws := GetOrCreateWordleState(chatID)
+	ws.Lock()
+	ws.Active = true
+	ws.Word = getRandomWordleWord()
+	ws.Guesses = make([]string, 0)
+	ws.Attempts = 0
+	ws.Unlock()
+
+	msg := fmt.Sprintf("🐊 🖼 *Wordle started!* ✨\n\n• The word consists of 5 letters.\n• You have %d attempts.\n\n💡 Hints:\n🟩 Correct letter in the right spot\n🟨 Correct letter but in the wrong spot\n🟥 Letter is not in the word\n\nSend a 5-letter word to guess.", ws.MaxAttempts)
+	view.SendMessage(bot, chatID, msg)
+}
+
+// HandleGuess processes a guess for a Wordle game
+func HandleGuess(bot *tgbotapi.BotAPI, message *tgbotapi.Message, client *mongo.Client, chatID int64, text string) {
+	ws := GetOrCreateWordleState(chatID)
+	ws.Lock()
+	defer ws.Unlock()
+
+	if !ws.Active {
+		return
+	}
+
+	guess := strings.ToLower(strings.TrimSpace(text))
+	if len(guess) != 5 {
+		return // Not a valid guess format, ignore
+	}
+
+	wordsMutex.RLock()
+	isValid := validWordleWords[guess]
+	wordsMutex.RUnlock()
+
+	if !isValid {
+		view.SendMessage(bot, chatID, fmt.Sprintf("❌ %s is not a valid word.", strings.ToUpper(guess)))
+		return
+	}
+
+	alreadyGuessed := false
+	for _, g := range ws.Guesses {
+		if g == guess {
+			alreadyGuessed = true
+			break
+		}
+	}
+
+	if alreadyGuessed {
+		view.SendMessage(bot, chatID, "⚠️ This word was already guessed!")
+		return
+	}
+
+	ws.Guesses = append(ws.Guesses, guess)
+	ws.Attempts++
+
+	board := buildWordleBoard(ws)
+
+	if guess == ws.Word {
+		ws.Active = false
+		msg := fmt.Sprintf("%s\n\n🟩 🟩 🟩 🟩 🟩  %s   [+25💎]\n🎉 [%s](tg://user?id=%d) guessed it in %d attempts!\n\nStart new Wordle! /wordle",
+			board, strings.ToUpper(ws.Word), message.From.FirstName, message.From.ID, ws.Attempts)
+
+		view.SendMessage(bot, chatID, msg)
+		go repository.InsertDoc(message.From.ID, message.From.FirstName, chatID, client, "CrocEn")
+	} else if ws.Attempts >= ws.MaxAttempts {
+		ws.Active = false
+		msg := fmt.Sprintf("%s\n\n❌ Out of attempts! The word was %s.\n\nStart new Wordle! /wordle", board, strings.ToUpper(ws.Word))
+		view.SendMessage(bot, chatID, msg)
+	} else {
+		view.SendMessage(bot, chatID, board)
+	}
+}
